LPS-84857 extract new class that handles the archetype framework and create a new ProjectGenerator that generates projects from archetypes. This refactoring also lays the groundwork for the new partial archetypes coming for code templates
