+++ modules/sdk/project-templates/project-templates/src/test/java/com/liferay/project/templates/ProjectTemplatesTest.java
@@ -17,7 +17,7 @@ package com.liferay.project.templates;
 import aQute.bnd.main.bnd;
 
 import com.liferay.maven.executor.MavenExecutor;
-import com.liferay.project.templates.internal.Archetyper;
+import com.liferay.project.templates.internal.ProjectGenerator;
 import com.liferay.project.templates.internal.util.FileUtil;
 import com.liferay.project.templates.internal.util.Validator;
 import com.liferay.project.templates.util.DirectoryComparator;
@@ -4014,7 +4014,7 @@ public class ProjectTemplatesTest {
 		projectTemplatesArgs.setService(service);
 		projectTemplatesArgs.setTemplate(template);
 
-		Archetyper archetyper = new Archetyper();
+		ProjectGenerator archetyper = new ProjectGenerator();
 
 		archetyper.generateProject(
 			projectTemplatesArgs, archetyperDestinationDir);
